% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi_sq.R
\name{assumptions_chi}
\alias{assumptions_chi}
\title{Assumptions_chi}
\usage{
assumptions_chi(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
graphical and numerical summary required to complete the test
}
\description{
Assumptions_chi
}
\examples{
assumptions_chi(project2022)
}
