% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{assumptions_linear}
\alias{assumptions_linear}
\title{Assumptions_linear}
\usage{
assumptions_linear(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
graphical and numerical summary required to complete the test
}
\description{
Assumptions_linear
}
\examples{
assumptions_linear(project2022)
}
