% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-test.R
\name{assumptions_t}
\alias{assumptions_t}
\title{Assumptions_t}
\usage{
assumptions_t(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
graphical and numerical summary required to complete the test
}
\description{
Assumptions_t
}
\examples{
assumptions_t(project2022)
}
