% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\name{check_changelog}
\alias{check_changelog}
\title{identify if the package contains a changelog}
\usage{
check_changelog(package_directory)
}
\arguments{
\item{package_directory}{the directory the package
source code lives in, returned from \code{\link{get_package_source}}}
}
\value{
TRUE if the package has something that looks
like a changelog, FALSE otherwise.
}
\description{
searches through the downloaded source code
of a package trying to identify if it has something that
looks like a log of changes between versions.
}

