% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_practices.R
\name{check_practices}
\alias{check_practices}
\title{Check all the practices of a package on CRAN}
\usage{
check_practices(package_name, metadata_lst = NULL, src_dir = NULL,
  error = TRUE)
}
\arguments{
\item{package_name}{character vector with the names of one or more packages
on CRAN}

\item{metadata_lst}{List of package metadata objects}

\item{src_dir}{If given, a directory that holds sources of the packages}

\item{error}{whether the function should raise an error if any individual
package does. If FALSE, shows the error message but continues}
}
\value{
A tbl_df with one row for each package, and columns describing
the practices followed by each.
}
\description{
Given the name of a package on CRAN, retrieve the source
and. Then run all tests in the practice package, returning a one-row
data frame with information on the package.
}
\details{
This relies entirely on other functions in the practice package.
}
\examples{
\dontrun{
check_practices("urltools")

# check multiple packages
practices <- check_practices(c("urltools", "ggplot2", "dplyr", "survival"))
practices

# extract what packages each links to
library(dplyr)
library(tidyr)

practices \%>\%
  select(package, links_to) \%>\%
  unnest(links_to)
}
}

