% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\name{check_roxygen}
\alias{check_roxygen}
\title{check whether a package's documentation uses roxygen2}
\usage{
check_roxygen(package_directory)
}
\arguments{
\item{package_directory}{the full path to the directory, retrieved
with \code{\link{get_package_source}}.}
}
\description{
identifies if a package's documentation uses
roxygen2's inline documentation, by scanning the R files
hunting for roxygen2 tags.
}

