% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\name{check_versioning}
\alias{check_versioning}
\title{identify if a package is semantically versioned}
\usage{
check_versioning(package_metadata)
}
\arguments{
\item{package_metadata}{package metadata retrieved with \code{\link{get_package_metadata}}.}
}
\value{
TRUE if the package does follow the semantic versioning standard,
FALSE if not.
}
\description{
take a package's metadata and identify
from it whether the package follows the "semantic versioning"
convention, at least in theory. See details for, well,
details.
}
\details{
\href{http://semver.org/}{semantic versioning}
is a convention for identifying version numbers of a codebase
in a way that distinguishes major changes, minor changes
and patches by using three period-separated
groups of digits. As an example, 2.0.0 is semantically
versioned; 2.0 is not.
}
\examples{
#Identify if ggplot2 is semantically versioned (it is)
ggplot_metadata <- get_package_metadata("ggplot2")
check_versioning(ggplot_metadata)

#Identify if A3 is semantically versioned (it isn't)
}

