% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/documentation.R
\name{check_vignettes}
\alias{check_vignettes}
\title{identify the presence and source of package
vignettes.}
\usage{
check_vignettes(package_directory)
}
\arguments{
\item{package_directory}{the directory the package
source code lives in, returned from \code{\link{get_package_source}}}
}
\value{
"No vignettes" if no vignettes are found, "Sweave" if
traditional, Sweave-formatted vignettes are found, and "Knitr"
if knitr-formatted vignettes are found.
}
\description{
identify how a package's vignettes are built, if it has
vignettes.
}
\details{
identifying how vignettes are created or stored isn't
trivial. One heuristic we can use is the actual storage format.
Traditional vignettes have been stored as ".Rnw" files, containing
both code and text. More recently, R Markdown files (".Rmd") have
become a preferred storage format, and are used by the Knitr vignette
building package by default.

\code{check_vignettes} hunts for a "vignettes" folder: if it
isn't found, there are no vignettes. If it is and contains files
with a ".Rnw" ending, there are Sweave vignettes. If there are
files with a ".Rmd" ending, Knitr. One edge case is the situation
where a package contains vignettes of both formats: the package's
vignettes would be identified as Knitr-based. Whether this
actually exists in the wild is not yet known.
}
\examples{
\dontrun{
#Identify how urltools vignettes are built, if it has any.
#It does, and if you run this code, you'll see the answer is
#"Knitr".
file_location <- get_package_source("urltools")
check_vignettes(file_location)

#Fortunes uses sweave
file_location <- get_package_source("fortunes")
check_vignettes(file_location)

#A3 has no vignettes
file_location <- get_package_source("A3")
check_vignettes(file_location)
}
}
\seealso{
\code{\link{check_roxygen}} to identify if a package's documentation is
formatted using roxygen2, \code{\link{check_changelog}} to see if there
is something that looks like a changelog, and the package index for more
tests and checks.
}

