% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi_sq.R
\name{conclusion_chi}
\alias{conclusion_chi}
\title{Conclusion_chi}
\usage{
conclusion_chi(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
the answer to your research question based on your decision
}
\description{
Conclusion_chi
}
\examples{
conclusion_chi(project2022)
}
