% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{conclusion_linear}
\alias{conclusion_linear}
\title{Conclusion_linear}
\usage{
conclusion_linear(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
the answer to your research question based on your decision
}
\description{
Conclusion_linear
}
\examples{
conclusion_linear(project2022)
}
