% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-test.R
\name{conclusion_t}
\alias{conclusion_t}
\title{Conclusion_t}
\usage{
conclusion_t(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
the answer to your research question based on your decision
}
\description{
Conclusion_t
}
