% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ecoregions.R, R/plot_ecoregions.R
\name{count_ecoregions}
\alias{count_ecoregions}
\title{Count ecoregions occurences}
\usage{
count_ecoregions(data)

count_ecoregions(data)
}
\value{
This function returns a data.table which species names and number
of ecoregions (\code{data.table}).

This function returns a data.table which species names and number
of ecoregions (\code{data.table}).
}
\description{
This function count the number of ecoregions where species are present.

This function count the number of ecoregions where species are present.
}
