% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{CRANversions}
\alias{CRANversions}
\title{Version release dates of each package in CRAN}
\format{A tbl_df with one row for each version, and three columns:
\itemize{
  \item package. package name
  \item version. package version
  \item date. release date as a POSIXct object
}}
\description{
The release date of each version of each package.
}

