% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{createIntegerObject}
\alias{createIntegerObject}
\title{Create Integer Object (S3)}
\usage{
createIntegerObject(x)
}
\arguments{
\item{x}{integer to be created as an S3 object}
}
\value{
object the integer S3 object with attributes
}
\description{
Takes an integer to create an S3 object to save data about the integer inside the the integer object.
}
\section{Quick Reference}{

\preformatted{createIntegerObject(9) gives an object like
x {
  square: 81
  factors: 1, 3, 9
}}
}

\examples{
createIntegerObject(9)
}
