% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi_sq.R
\name{decision_chi}
\alias{decision_chi}
\title{Decision_chi}
\usage{
decision_chi(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
the decision of your test based on your p.value and a significance level of 5\%
}
\description{
Decision_chi
}
\examples{
decision_chi(project2022)
}
