% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{decision_linear}
\alias{decision_linear}
\title{Decision_linear}
\usage{
decision_linear(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
the decision of your test based on your p.value and a signifance level of 5\%
}
\description{
Decision_linear
}
\examples{
decision_linear(project2022)
}
