% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-test.R
\name{decision_t}
\alias{decision_t}
\title{Decision_t}
\usage{
decision_t(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
the decision of your test based on your p.value and a signifance level of 5\%
}
\description{
Decision_t
}
