% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{deparse2}
\alias{deparse2}
\title{Deparse 2 (Wrapper for deparse in pryr)}
\usage{
deparse2(express)
}
\arguments{
\item{express}{}
}
\value{
a string representation of the expression
}
\description{
Takes an expression and deparses it into a string. Unlike the original deparse, it will always be one string.
}
\section{Quick Reference}{

\preformatted{deparse2(1 + 8) outputs "1 + 8"
deparse2(1:10) outputs 1:10
deparse2(a + b + c + d + e + ... + z) outputs a + b + ... + z }
}

\examples{
deparse2(1 + 8)
deparse2(1:12)
deparse2(square(17))
}
