% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{divisors_of}
\alias{divisors_of}
\title{Divisors Of}
\usage{
divisors_of(x)
}
\arguments{
\item{x}{the number to find the divisors of}
}
\value{
Returns a vector of divisors, must include 1 and the x itself if x was a single integer.
Returns a list of vectors if x was a vector of integers
}
\description{
Divisors Of
}
\note{
This function is not exactly vectorized as passing a vector returns a list of vectors
}
\examples{
divisors_of(12) outputs [ 1, 2, 3, 4, 6, 12 ]
divisors_of(7) outputs  [ 1, 7 ]
divisors_of(c(10, 14, 5)) outputs list( [ 1, 2, 5, 10], [ 1, 2, 7, 14 ], [ 1, 5 ] )
}
