% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_wildfinder_data.R
\name{dl_wildfinder_data}
\alias{dl_wildfinder_data}
\title{Download Wildfinder datasets}
\usage{
dl_wildfinder_data()
}
\value{
This function returns the path (\code{character}) to the downloaded files
(e.g. \code{data/wildfinder/wildfinder-ecoregions_list.csv}).
}
\description{
This function downloads the Wildfinder datasets (csv files) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}.
The files \code{wildfinder-ecoregions_list.csv}, \code{wildfinder-ecoregions_species.csv}
and \code{wildfinder-mammals_list.csv} will be stored in #' \verb{data/wildfinder/}.
This folder will be created if required.
}
