% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_all_CRAN_metadata}
\alias{download_all_CRAN_metadata}
\title{Download CRAN metadata, including all versions of each package}
\usage{
download_all_CRAN_metadata(outfile, ...)
}
\arguments{
\item{outfile}{output .rdata file}

\item{...}{extra arguments passed on to \code{download_packages_metadata},
such as \code{verbose}}
}
\value{
The CRANmetadata object, invisible
}
\description{
Download all metadata from all CRAN packages, optionally saving it to a file.
}

