% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_package.R
\name{download_packages}
\alias{download_packages}
\title{Download multiple package sources from CRAN.}
\usage{
download_packages(package_names, dir, quiet = TRUE)
}
\arguments{
\item{package_names}{character vector of one or more package names}

\item{dir}{directory to save packages to}

\item{quiet}{whether package download messages should be suppressed}
}
\description{
Download multiple package sources from CRAN, saving them to a
directory. This can be used in a meta-analysis of CRAN.
}

