% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_packages_metadata}
\alias{download_packages_metadata}
\title{Download metadata from multiple packages to a file}
\usage{
download_packages_metadata(package_names, verbose = FALSE, ...)
}
\arguments{
\item{package_names}{Character vector of package names; retrievable with
\code{\link{get_package_names}}}

\item{verbose}{Whether package names should be displayed as they are
downloaded}

\item{...}{Extra arguments passed on to \code{\link{get_package_metadata}}}
}
\value{
A named list, one component for each package, containing a list of
  metadata for the package. This list contains the contents from the
  \href{https://github.com/metacran/crandb}{CRAN API}, with one field added:
  \code{retrieved}, a \code{\link[=DateTimeClasses]{POSIXct}} object with
  the time that the data was retrieved.
}
\description{
Download metadata from multiple packages to a file. This is useful for larger
analyses of CRAN.
}
\examples{
foo <- download_packages_metadata(c("broom", "tidyr"))
str(foo, max.level = 2)
}

