% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dates.R
\name{earliest_release}
\alias{earliest_release}
\title{Identify the Date of the First Release}
\usage{
earliest_release(metadata)
}
\arguments{
\item{metadata}{a metadata object returned from \code{\link{get_package_metadata}}}
}
\description{
Takes package metadata obtained with \code{\link{get_package_metadata}}
and returns a date object containing the date of the first CRAN package release.
}
\examples{
\dontrun{
package_metadata <- get_package_metadata("abf2")
earliest_release(package_metadata)
}
}

