% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fastfib}
\alias{fastfib}
\title{Fast Fibonacci Function}
\usage{
fastfib(x)
}
\arguments{
\item{x}{the index of the desired fibonacci number}
}
\value{
vector of the fibonacci number of the indeces given
}
\description{
Fast Fibonacci executes in constant time by using a formula instead of recursion
}
\note{
This function is vectorized
}
\section{Quick Reference}{

\preformatted{fastfib(2) outputs 1
fastfib(4) outputs 3
fastfib(c(1, 5, 8)) outputs [ 1, 5, 21 ]}
}

\examples{
fastfib(1)
fastfib(2)
fastfib(4)
fastfib(c(1, 5, 8))
}
