% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Merge two factors into one}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor (or character vector)}

\item{b}{factor (or charcter vector)}
}
\value{
factor
}
\description{
Merge two factors into one
}
