% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcount.R
\name{fcount}
\alias{fcount}
\title{Count the frequency of each factor element}
\usage{
fcount(x)
}
\arguments{
\item{x}{factor vector}
}
\value{
frequency table (matrix)
}
\description{
Count the frequency of each factor element
}
\examples{
fcount(as.factor("a", "b", "c", "c", "d", "d", "d", "d"))
}
