% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fget}
\alias{fget}
\title{Function Getter}
\usage{
fget(name, env, inherits = TRUE)
}
\arguments{
\item{name}{the name of the function to search for}

\item{env}{the starting environment}

\item{inherits}{flag for allowing the function to look in parent environments also}
}
\value{
a function
}
\description{
fget() returns a list of th
}
\examples{
fget("ls", environment())
}
