% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi_sq.R
\name{fit_chi}
\alias{fit_chi}
\title{Fit_chi}
\usage{
fit_chi(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
summary of your important values to carry out the test
}
\description{
Fit_chi
}
\examples{
fit_chi(project2022)
}
