% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{fit_linear}
\alias{fit_linear}
\title{Fit_linear}
\usage{
fit_linear(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
summary of your important values to carry out the test
}
\description{
Fit_linear
}
\examples{
fit_linear(project2022)
}
