% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-test.R
\name{fit_t}
\alias{fit_t}
\title{Fit_t}
\usage{
fit_t(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
summary of your important values to carry out the test
}
\description{
Fit_t
}
