% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/license.R
\name{get_license}
\alias{get_license}
\title{extract the licenses used by a package}
\usage{
get_license(package_metadata)
}
\arguments{
\item{package_metadata}{package metadata retrieved with \code{\link{get_package_metadata}},
or the name of a package (in which case \code{get_package_metadata} will be called
internally).}
}
\value{
a character vector containing each unique license mentioned in the
package's DESCRIPTION field.
}
\description{
when provided with package metadata, \code{get_license}
extracts each unique license a package is released under and provides
them as a vector.
}
\details{
license information is not stored in a tremendously consistent way, but
we can make a best-guess at it by stripping out "+ file license" strings, and
then splitting on |.
}
\examples{
#What license(s) are urltools under?
get_license(get_package_metadata("urltools"))
}
\seealso{
the package index for more checks and functionality.
}

