% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{get_package_downloads}
\alias{get_package_downloads}
\title{retrieve the number of downloads of a package}
\usage{
get_package_downloads(package, start_date, end_date, all = FALSE)
}
\arguments{
\item{package}{the name of a package.}

\item{start_date}{the start of the date range to cover.}

\item{end_date}{the end of the date range to cover.}

\item{all}{whether to retrieve data for all packages,
rather than one specific package. Set to FALSE by
default.}
}
\description{
this identifies the number of downloads
of packages from the RStudio CRAN mirror; it can be
used as a measure of package popularity and,
by extension, a measure of the importance of a particular
package having good engineering standards
}

