% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_package.R
\name{get_package_source}
\alias{get_package_source}
\title{Retrieve a package source}
\usage{
get_package_source(package_name)
}
\arguments{
\item{package_name}{the name of a package.}
}
\value{
a full, non-relative link to the directory the decompressed
package is stored in.
}
\description{
\code{get_package_source} retrieves a package source, decompresses
it, and places it in a directory so that software engineering tests can be performed
on the contents. See the "See also" section for examples of these tests, and
\code{remove_package_source} to remove the source directory when you're done
using it.
}
\seealso{
\code{\link{remove_package_source}} to remove the package, \code{\link{check_vignettes}}
to
}

