% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{heat_map}
\alias{heat_map}
\title{A heatmap function}
\usage{
heat_map(df)
}
\arguments{
\item{d}{dataframe}
}
\description{
This function will create a heatmap for rows in a data frame
}
\examples{
heat_map(d = USArrests)
}
