% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi_sq.R
\name{hypothesis_chi}
\alias{hypothesis_chi}
\title{Hypothesis_chi}
\usage{
hypothesis_chi(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
the null and alternate hypothesis
}
\description{
Hypothesis_chi
}
\examples{
hypothesis_chi(project2022)
}
