% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{hypothesis_linear}
\alias{hypothesis_linear}
\title{Hypothesis_linear}
\usage{
hypothesis_linear(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
the null and alternate hypothesis
}
\description{
Hypothesis_linear
}
\examples{
 hypothesis_linear(project2022)
}
