% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-test.R
\name{hypothesis_t}
\alias{hypothesis_t}
\title{Hypothesis_t}
\usage{
hypothesis_t(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
the null and alternate hypothesis
}
\description{
Hypothesis_t
}
\examples{
hypothesis_t(project2022)
}
