% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_tables.R
\name{join_tables}
\alias{join_tables}
\title{Join tables}
\usage{
join_tables(data_ursidae, data_ecoregions, data_ecor_spec)
}
\value{
This function returns a data.table which contains all informations
on the Usidae family (\code{data.table}).
}
\description{
This function join the tables on Ursidae data.
}
