% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_practices.R
\name{make_vector_list}
\alias{make_vector_list}
\title{If a vector is of length greater than 1, make it a list;
if it's NULL, return NA}
\usage{
make_vector_list(x)
}
\arguments{
\item{x}{a vector}
}
\description{
This is useful for creating a data frame that may have
list-columns
}

