% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{my_sample}
\alias{my_sample}
\title{입력값의 제곱을 구하는 함수}
\usage{
my_sample(x)
}
\arguments{
\item{x}{입력값의 형태와 자세한 설명을 해줍니다. 정수를 입력하세요~}
}
\value{
결과 값에 대한 자세한 설명을 합니다. x의 제곱 값이 나옴!
}
\description{
입력값의 제곱을 구하는 함수
}
