% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi_sq.R
\name{mytest_chi}
\alias{mytest_chi}
\title{Mytest_chi}
\usage{
mytest_chi(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
all functions outputted into a nicely formatted report
}
\description{
Mytest_chi
}
\examples{
mytest_chi(project2022)
}
