% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{mytest_linear}
\alias{mytest_linear}
\title{Mytest_linear}
\usage{
mytest_linear(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
all functions outputted into a nicely formatted report
}
\description{
Mytest_linear
}
