% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-test.R
\name{mytest_t}
\alias{mytest_t}
\title{Mytest_t}
\usage{
mytest_t(filename)
}
\arguments{
\item{filename}{dataset that you want to be tested}
}
\value{
all functions outputted into a nicely formatted report
}
\description{
Mytest_t
}
