% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{optimal_cluster}
\alias{optimal_cluster}
\title{optimal_cluster}
\usage{
optimal_cluster(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
This function creates an elbow plow, silhouette,
and gap statistic to help determine the optimal number of clusters
for a k-means cluster anaylsis
}
\examples{
optimal_cluster(df = iris[, -5])
}
