% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-package.R
\docType{package}
\name{practice-package}
\alias{practice}
\alias{practice-package}
\title{practice: optimal_cluster}
\description{
This function creates an elbow plow, silhouette, and gap statistic to help determine the optimal number of clusters for a k-means cluster anaylsis
}
\keyword{internal}
