% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/practice.R
\docType{package}
\name{practice}
\alias{practice}
\alias{practice-package}
\title{Tools for evaluating how a package adheres to common software engineering practices}
\description{
R has unit tests, documentation suites,
integration with upstream repositories for bug reports, and
a whole lot more - but not a lot of people actually use them.
}

