% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project2022}
\alias{project2022}
\title{project2022 dataset}
\format{
\subsection{\code{project2022}}{

A data frame with 1,000 rows and 5 columns:
\describe{
\item{ID}{ID of sample individual}
\item{gender}{male or female}
\item{height}{in cm}
\item{weight}{in kg}
\item{phys}{amount of physical activity}

...
}
}
}
\source{
STAT1378
}
\usage{
project2022
}
\description{
Contains 1000 observations of men and women aged 26-45 with information on ID,
gender, weight, height and amount of physical activity
}
\keyword{datasets}
