% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ursidae.R
\name{select_ursidae}
\alias{select_ursidae}
\title{Select Usidae data}
\usage{
select_ursidae(data)
}
\value{
This function returns a data.table which contains only species
from the Usidae family (\code{data.table}).
}
\description{
This function select only Ursidae data the Wildfinder species dataset.
}
