% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplesum.R
\name{simplesum}
\alias{simplesum}
\title{yoyoyo}
\usage{
simplesum(x, y)
}
\arguments{
\item{x, y}{two values to be summed}
}
\value{
the sum of x and y
}
\description{
check this out
}
\examples{
simplesum(2,3)
}
