% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{square}
\alias{square}
\title{Square Function}
\usage{
square(x)
}
\arguments{
\item{x}{the vector to be squared}
}
\value{
The vector with all elements squared
}
\description{
Square Function
}
\note{
: This function is vectorized
}
\examples{
square(9) outputs 81
square(-7) outputs 49
square(c(1, 5, 8)) outputs [ 1 25 64 ]
}
