% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.foo.R
\name{test.foo}
\alias{test.foo}
\title{mean of some vector of numbers}
\usage{
test.foo(x)
}
\arguments{
\item{x}{A vector of numbers.}
}
\value{
The mean of numbers in \code{x}.
}
\description{
mean of some vector of numbers
}
\examples{
test.foo(c(1, 9))
}
