% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{update_CRAN_practices}
\alias{update_CRAN_practices}
\title{Update the CRAN practices dataset}
\usage{
update_CRAN_practices(data_dir = "data", src_dir = NULL, ...)
}
\arguments{
\item{data_dir}{Output directory to write \code{CRANpractices.rda}}

\item{src_dir}{Directory of package sources}

\item{...}{extra arguments passed on to \code{\link{check_practices}}}
}
\description{
Check all CRAN practices, and save them to a directory.
}
\details{
This works with the current version of the \code{\link{CRANmetadata
  dataset}} (make sure you've re-built the package if you've updated that
  dataset).
}

