% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{update_CRAN_versions}
\alias{update_CRAN_versions}
\title{Given a list with versions metadata from CRAN, create
the CRANversions dataset}
\usage{
update_CRAN_versions(all_metadata, data_dir = "data")
}
\arguments{
\item{all_metadata}{Named list of metadata for CRAN packages,
including all versions}

\item{data_dir}{directory to save the dataset to}
}
\value{
The CRANversions dataset, invisibly
}
\description{
Given a list with versions metadata from CRAN, create
the CRANversions dataset
}
\examples{
\dontrun{

# this line may take 5-10 minutes
CRANallmetadata <- practice:::download_all_CRAN_metadata("all_metadata.rdata", verbose = TRUE)

update_CRAN_versions(CRANallmetadata)

}
}

