% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{visualization}
\usage{
visualization(d, n)
}
\arguments{
\item{d}{dataframe}

\item{n}{number of clusters}
}
\description{
This function creates plots to visualize data for a number of
clusters
}
\examples{
visualization(d = iris, n = 2:4)
}
\keyword{cluster}
\keyword{k-means}
