% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfinder_read.R
\name{wildfinder_read}
\alias{wildfinder_read}
\title{Read Wildfinder datasets}
\usage{
wildfinder_read()
}
\value{
This function returns a list containing the datasets (\code{tibble}).
}
\description{
This function read the Wildfinder data files (csv files).
}
