% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\name{check_changelog}
\alias{check_changelog}
\title{identify if the package contains a changelog}
\usage{
check_changelog(package_directory)
}
\arguments{
\item{package_directory}{the directory the package
source code lives in, returned from \code{\link{get_package_source}}}
}
\value{
TRUE if the package has something that looks
like a changelog, FALSE otherwise.
}
\description{
searches through the downloaded source code
of a package trying to identify if it has something that
looks like a log of changes between versions.
}
\details{
changelogs - logs of what changes in a package
between versions, be that bug fixes, new features or alterations
to how existing features function - are common to R packages
and collections of code generally.

With R packages, they are normally stored as "NEWS"
or "CHANGELOG" files. \code{check_changelog} checks
for the presence of a changelog by seeing if there are any files
in the root directory of the package source code that match
the case-insensitive regular expression "(change|news)".
}
\examples{
\dontrun{
#Does A3 have a changelog? (yes)
file_location <- get_package_source("A3")
result <- check_changelog(file_location)

#What about "fortunes"? (no)
file_location <- get_package_source("fortunes")
result <- check_changelog(file_location)
}
}

