% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/casing.R
\name{check_package_naming}
\alias{check_package_naming}
\title{classify the format of a package's name}
\usage{
check_package_naming(package_metadata)
}
\arguments{
\item{package_metadata}{package metadata retrieved with \code{\link{get_package_metadata}}}
}
\value{
a named, two-element list containing "casing" ("Lower", "Upper" or "Mixed")
and "alphanumeric" (TRUE for package names that exclusively contain alphanumeric characters,
and FALSE for package names that contain special characters.)
}
\description{
package names come likethis and like.this
and Like.THIS and in a whole host of formats. \code{check_package_naming}
consumes package metadata and identifies the format a package's name follows,
using the convention described below.
}

