% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\name{check_roxygen}
\alias{check_roxygen}
\title{check whether a package's documentation uses roxygen2}
\usage{
check_roxygen(package_directory)
}
\arguments{
\item{package_directory}{the full path to the directory, retrieved
with \code{\link{get_package_source}}.}
}
\description{
identifies if a package's documentation uses
roxygen2's inline documentation, by scanning the R files
hunting for roxygen2 tags.
}
\details{
checking for roxygen2 is fairly simple. One way is to
look at the ./man files and search for the "% Generated by roxygen2"
tag, but early versions of roxygen2 didn't introduce that tag.
Instead, we scan the .R files hunting for actual tags, specifically
searching for @export (which every package, presumably, must have
at least once - there wouldn't actually be any R it provides if
not.)
}
\examples{
#Identify if ggplot2 uses roxygen
file_location <- get_package_source("ggplot2")
check_roxygen(file_location)
remove_package_source(file_location)
}

