% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testing.R
\name{check_testing}
\alias{check_testing}
\title{identify a package's unit testing framework, if it has one}
\usage{
check_testing(package_directory)
}
\arguments{
\item{package_directory}{the full path to the directory, retrieved
with \code{\link{get_package_source}}.}
}
\value{
either "None" (no recogniseable tests), "RUnit" (RUnit-based
tests), "testthat" (testthat-based tests) or "Other" (some form
of hand-rolled testing).
}
\description{
takes a downloaded package and identifies if the package
has a unit testing framework, or anything indicative of automated testing.
}

