% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testing.R
\name{check_upstream_repository}
\alias{check_upstream_repository}
\title{identifies if a package indicates it has a versioned repository for
upstream patches or bug reports.}
\usage{
check_upstream_repository(package_metadata)
}
\arguments{
\item{package_metadata}{package metadata retrieved with \code{\link{get_package_metadata}}}
}
\value{
"GitHub", "R-Forge", "BitBucket", "Email" or "None/Other".
}
\description{
A crucial element of good software is software that users can patch
(or at least complain about). \code{check_upstream_repository} consumes package
metadata and identifies from it if users are provided with an actual versioning
system, or contact email, to forward issues or fixes to.
}

