% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\name{check_vignettes}
\alias{check_vignettes}
\title{identify the presence and source of package
vignettes.}
\usage{
check_vignettes(package_directory)
}
\arguments{
\item{package_directory}{the directory the package
source code lives in, returned from \code{\link{get_package_source}}}
}
\value{
"No vignettes" if no vignettes are found, "Sweave" if
traditional, Sweave-built vignettes are found, and "Knitr"
if knitr-built vignettes are found.
}
\description{
identify how a package's vignettes are built, if it has
vignettes.
}

