% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/documentation.R
\name{check_vignettes}
\alias{check_vignettes}
\title{identify the presence and source of package
vignettes.}
\usage{
check_vignettes(package_directory)
}
\arguments{
\item{package_directory}{the directory the package
source code lives in, returned from \code{\link{get_package_source}}}
}
\value{
A list with two components:
  \item{Format}{"None" if no vignettes are found, "LaTeX" if
LaTeX .Rnw files are found, "Markdown" if .Rmd files are found.}
  \item{Builder}{The package used to compile the vignettes,
typically Sweave or knitr.}
}
\description{
identify how a package's vignettes are built, if it has
vignettes.
}
\details{
There are two things to look for in terms of vignettes: the format,
either LaTeX or R markdown, and the vignette builder, typically
Sweave or knitr. All .Rmd files are knitr, but .Rnw can be compiled by
either.

\code{check_vignettes} looks for a \code{vignettes}
folder to see whether .Rnw (LaTeX) files or .Rmd (markdown) files are
found. It also checks for a VignetteBuilder option in the DESCRIPTION.
}
\examples{
\dontrun{
#Identify how urltools vignettes are built, if it has any.
#It does, and if you run this code, you'll see the answer is
#"Knitr".
file_location <- get_package_source("urltools")
check_vignettes(file_location)

#Fortunes uses sweave
file_location <- get_package_source("fortunes")
check_vignettes(file_location)

#A3 has no vignettes
file_location <- get_package_source("A3")
check_vignettes(file_location)
}
}
\seealso{
\code{\link{check_roxygen}} to identify if a package's documentation is
formatted using roxygen2, \code{\link{check_changelog}} to see if there
is something that looks like a changelog, and the package index for more
tests and checks.
}

