% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_packages_metadata}
\alias{download_packages_metadata}
\title{Download metadata from multiple packages to a file}
\usage{
download_packages_metadata(package_names, verbose = FALSE, ...)
}
\arguments{
\item{package_names}{Character vector of package names}

\item{verbose}{Whether package names should be displayed as they are downloaded}

\item{...}{Extra arguments passed on to \code{get_package_metadata}}
}
\value{
A named list, one for each package, containing a list for each package
metadata. This list contains the contents from the CRAN API, with one
field added: \code{retrieved}, a \code{posixCT} object with the time that the
data was retrieved.
}
\description{
Download metadata from multiple packages to a file. This is useful for
larger analyses of CRAN.
}

