% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{factorial2}
\alias{factorial2}
\title{Factorial Function}
\usage{
factorial2(x)
}
\arguments{
\item{x}{the object to be factorialized}
}
\value{
the resulting vector of factorialized numbers
}
\description{
Factorial Function
}
\note{
This function is vectorized
}
\examples{
factorial2(2) outputs 4
factorial2(4) outputs 24
factorial2(c(0, 3, 5)) outputs [ 1 6 120 ]
}
