% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fastfib}
\alias{fastfib}
\title{Fast Fibonacci Function}
\usage{
fastfib(x)
}
\arguments{
\item{x}{the index of the desired fibonacci number}
}
\value{
vector of the fibonacci number of the indeces given
}
\description{
Fast Fibonacci executes in constant time by using a formula instead of recursion
}
\note{
This function is vectorized
}
\examples{
fib(2) outputs 1
fib(4) outputs 3
fib(c(1, 5, 8)) outputs [ 1 5 21 ]
}
