% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fib}
\alias{fib}
\title{Fibonacci Function}
\usage{
fib(x)
}
\arguments{
\item{x}{the index of the desired fibonacci number}
}
\value{
vector of the fibonacci number of the indeces given
}
\description{
The fib function calculates the fibonacci number at the given index
}
\note{
This function is vectorized
}
\section{Quick Reference}{

\preformatted{fib(2) outputs 1
fib(4) outputs 3
fib(c(1, 5, 8)) outputs [ 1, 5, 21 ]}
}

\examples{
fib(1)
fib(2)
fib(4)
}
