% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/license.R
\name{get_license}
\alias{get_license}
\title{extract the licenses used by a package}
\usage{
get_license(package_metadata)
}
\arguments{
\item{package_metadata}{package metadata retrieved with \code{\link{get_package_metadata}}.}
}
\value{
a character vector containing each unique license mentioned in the
package's DESCRIPTION field.
}
\description{
when provided with package metadata, \code{get_license}
extracts each unique license a package is released under and provides
them as a vector.
}

