% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{get_package_downloads}
\alias{get_package_downloads}
\title{retrieve the number of downloads of a package}
\usage{
get_package_downloads(packages = NULL, start_date, end_date, all = FALSE)
}
\arguments{
\item{packages}{the name of a package, or a vector of
multiple package names.}

\item{start_date}{the start of the date range to cover.}

\item{end_date}{the end of the date range to cover.}

\item{all}{whether to retrieve data for all packages,
rather than one specific package. Set to FALSE by
default.}
}
\description{
this identifies the number of downloads
of packages from the RStudio CRAN mirror; it can be
used as a measure of package popularity and,
by extension, a measure of the importance of a particular
package having good engineering standards
}
\examples{
#ggplot2 downloads in March 2015
get_package_downloads("ggplot2", "2015-03-01", "2015-03-31")

#all package downloads on 1-2nd April
get_package_downloads(NULL, "2015-04-01", "2015-04-02", all = TRUE)
}

