% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_package.R
\name{get_package_metadata}
\alias{get_package_metadata}
\title{get metadata associated with a package on CRAN}
\usage{
get_package_metadata(package_name)
}
\arguments{
\item{package_name}{the name of a package, which can be retrieved with
\code{\link{get_package_names}}}
}
\description{
pings the \href{http://crandb.r-pkg.org/}{crandb} CRAN metadata
service to retrieve metadata associated with a specific package.
}

