% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/links.R
\name{links_from}
\alias{links_from}
\title{Get a count of packages that mention a specific package}
\usage{
links_from(package_metadata)
}
\arguments{
\item{package_metadata}{package metadata retrieved with \code{\link{get_package_metadata}}}
}
\value{
a one-element, numeric vector containing a count of how many packages depend
on the package \code{package_metadata} covers.
}
\description{
consumes packageX's metadata and returns a count
of how many packages depend on it.
}

