% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-package.R
\docType{package}
\name{practice-package}
\alias{practice}
\alias{practice-package}
\title{practice: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JoleneSutton/practice}
  \item Report bugs at \url{https://github.com/JoleneSutton/practice/issues}
}

}
\author{
\strong{Maintainer}: Jolene Sutton \email{jolene.sutton@gmail.com}

}
\keyword{internal}
