% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-package.R
\docType{package}
\name{practice-package}
\alias{practice}
\alias{practice-package}
\title{practice: practice}
\description{
This package aids in running K-means cluster analysis and principle component analysis on a data frame. {optimal_cluster} will help determine the optimal number of clusters in you data. {visulization} visualizes a number of clusters. Finall PCA will run and visualize a principle component analysis. The NBA and NBA_salary dataset is inculded to demonstrate these functions.
}
\keyword{internal}
