% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{square}
\alias{square}
\title{Square Function}
\usage{
square(x)
}
\arguments{
\item{x}{the vector to be squared}
}
\value{
The vector with all elements squared
}
\description{
the square function raises x to the second power
}
\note{
: This function is vectorized
}
\section{Quick Reference}{

\preformatted{square(9) outputs 81
square(-7) outputs 49
square(c(1, 5, 8)) outputs [ 1 25 64 ]}
}

\examples{
square(9)
square(-7)
square(c(1, 5, 8))
}
