% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{update_CRAN_metadata}
\alias{update_CRAN_metadata}
\title{Update the CRAN metadata dataset}
\usage{
update_CRAN_metadata(data_dir = "data", all = FALSE, ...)
}
\arguments{
\item{data_dir}{Output directory to write CRANmetadata.rda}

\item{all}{Whether all versions should be downloaded, or just the latest version
(by default, only latest version)}

\item{...}{Extra arguments passed on to \code{download_packages_metadata}}
}
\value{
The CRANmetadata object, invisible
}
\description{
Download metadata from all CRAN packages, and save it into a directory
(by default, \code{data}). This is used to create the \code{CRANmetdata}
dataset.
}

